data_fetching/requirements.txt
python-dotenv
requests
beautifulsoup4
google-cloud-storage
eventregistry
gcp_key.json